A Pipeable Wrapper For dbscan

A Pipeable Wrapper For dbscan

Usage

dbscan(data, eps, minPts, borderPoints = TRUE)

Arguments

data

A data.frame.

eps

size (radius) of the epsilon neighborhood. Can be omitted if x is a frNN object.

minPts

number of minimum points required in the eps neighborhood for core points (including the point itself).

borderPoints

logical; should border points be assigned to clusters. The default is TRUE for regular DBSCAN. If FALSE then border points are considered noise (see DBSCAN* in Campello et al, 2013).

Value

dbscan() returns an object of class dbscan_fast with the following components:

eps

value of the eps parameter.

minPts

value of the minPts parameter.

metric

used distance metric.

cluster

A integer vector with cluster assignments. Zero indicates noise points.

is.corepoint() returns a logical vector indicating for each data point if it is a core point.

Details

Adds a new column named clust to the data with the cluster results