Perform a Function on Each Cell in a Matrix

This performs the function without changing the matrix, then applies all changes at once.

Usage

flash(m, .fn, i = 1, cores = future::availableCores() - 1, ...)

Arguments

m

A matrix

.fn

The function to apply

i

The number of times to apply the function to the whole matrix

cores

The amount of cores to use when applying the function to each cell

Value

A matrix of the same dimensions as m