follow_directions

follow_directions

Usage

follow_directions(.data, row = 0, col = 0, preserve_data = FALSE)

Arguments

.data

A vector of RLUD directions. See expand_directions to generate from a data.frame of direction and magnitude.

row

(Default: 0) The starting row value

col

(Default: 0) The starting column value

preserve_data

A boolean to indicate if data besides row and col should be returned.

Value

A data.frame of coordinate pairs with row and col values.